fix: rename spiral
